requirements.txt
numpy
astropy-helpers
matplotlib
ads
synphot
astroquery
